chore(refactoring): move vital modules to components
* separate reply markup
* separate control flow components
