+++ src/components/control-flow/separation/index.ts
@@ -1,0 +1,1 @@
+export { Separation } from "./separation";
