+++ src/components/control-flow/types.ts
@@ -1,0 +1,6 @@
+import { UserId } from "@/types/user";
+import { Context } from "grammy";
+
+export interface IControlFlow {
+  sendData: (ctx: Context, userId: UserId) => void,
+}
