+++ src/scripts/handlers/locationButtons.ts
@@ -1,8 +1,8 @@
+import { removeInlineKeyboard } from "@/components/reply-markup";
 import { storeController } from "@/env";
 import { LocationsList } from "@/types/location";
 import { ButtonClickHandler } from "@/types/mixins";
 import { UserStatus } from "@/types/user";
-import { removeInlineKeyboard } from "@/utils/keyboard";
 import { initUserForExcursion } from "../tabs/excursion";
 
 export const locationButtonClickHandler: ButtonClickHandler = async ({ctx, userId, clickData}) => {
